pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trip-planner-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-adk>=1.26.0",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.41.0",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=24.0.0",
    "rich>=13.7.0",
    "bcrypt>=5.0.0",
    "flights>=0.7.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths      = ["tests"]
pythonpath     = ["src", "tests"]
markers = [
    "L1: Component level unit tests",
    "L2: Trajectory / integration evaluation tests (real LLM calls)",
]
asyncio_mode = "auto"


[tool.ruff]
line-length    = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version      = "3.12"
strict              = true
ignore_missing_imports = true

[project.scripts]
trip-planner     = "cli.main:app"
trip-planner-api = "api.app:start"
trip-planner-web = "cli.main:web"
